refactored community.jsx
